Improve ref check of constructors
In typer, self-invocation may be nested in a block
due to named arg rewrite. Check the result expr for
a forward self-invocation.

In ForwardDepChecks, the rhs of the ctor def
may be an expr (not a block) due to a forwarder
created by ElimRepeated. That means prepareStats
is not called before checking the Apply and there
is no current LevelInfo.

For checking, that forwarder gets a synthetic span.