Don't absolutize sources outside of the sourceroot in TASTy
With inlining and standard library patching, it's inevitable that, when
writing TASTy files, we'll encounter sources outside the sourceroot.
However, we shouldn't write those sources as absolute paths because
that's non-reproducible and can cause determinism issues. Instead, we
should write the paths as-is.

While these sorts of paths aren't correct when written relatively, they
weren't correct to begin with because they don't actually exist on the
filesystem.