Fix compiler crash in PatternMatcher
My understanding is that the test case in i24504 is expected to take
this branch:
https://github.com/scala/scala3/blob/0470ec1ff46a822ca58e204587fb67d56004623a/compiler/src/dotty/tools/dotc/transform/PatternMatcher.scala#L420
where we the info is comming from
https://github.com/scala/scala3/blob/0470ec1ff46a822ca58e204587fb67d56004623a/compiler/src/dotty/tools/dotc/transform/PatternMatcher.scala#L118-L124
that has called `dropNamedTuple`
https://github.com/scala/scala3/blob/0470ec1ff46a822ca58e204587fb67d56004623a/compiler/src/dotty/tools/dotc/transform/PatternMatcher.scala#L113-L115

the problem is that under this case the type will be `NamedTuple.From`
for which `isNamedTupleType` returns false. Which means that the
returned type will not be a `NonEmptyTupleType`.

The fix in this mr adds a `normalized.dealias` call in
`isNormalizedTupleType` which brings it inline with
`stripNormalizedTupleType`. But seems feasible that a better fix could
be that `NamedTuple.From` should have already been removed earlier. But
it not clear to me if that possible.

fixes: i24504