Somewhat better error notes for box failures
At least we don't get

    Note that cap is not included in {}

anymore. To make this even better we'd have to trace back a box
failure to the original types we were trying to compare. Right now
the box failure error notes are too disconnected from the rest. But
they become clearer when one compiles with -explain.