Use a better span for an anonymous class
An anonymous class is expanded roughly to

    { class $anon{ ... }; new $anon }

Previously the `new` part had a zero extent span at the end of the class.
When used with significant indentation, this caused some puzzlement in error
messages since one thought the last class definition was to blame. The span
is now the whole anonymous class.