Tweak cc error messages
Sometimes we have more than one representation of the same symbol, i.e.
both `^` and `cap`, or both `=>` and `cap`. If the cap is rendered first, we expand
the following string to `^{cap}` or `->{cap}` in order make it clearer that
they are the same. Unfortunately it only works if the cap comes first.