Use a LazyKeySet for LinkedHashMap's keySet
LinkedHashMap's `keySet` has always been returning a LinkedKeySet, a private data structure that was extended from the
generic MapOps.KeySet. Recently MapOps.KeySet was changed to a strict data structure, so we change the `keySet` method to
use the lazy wrapper instead. Unfortunately we cannot change the hierarchy of `LinkedKeySet`, but the API never said we have to use it.