Also mark free underlying capture set of local mutable variables
When we have

  var x: A^{c} = ...

where `x` is a local variable then when dereferencing `x` we also need to charge `c`.
This was forgotten before. For fields it's not a problem since `c` would already have been
charged for the prefix `p` in `p.x`.

An alternative would be to consider mutable variables to have boxed types. Then the same
capture set would be charged on unboxing.