Handle MissingType during Erasure phase
Fixes https://github.com/scala/scala3/issues/20491

Wrap the `transform` method in Erasure with a try-catch to handle `MissingType`.
These errors can occur during TypeErasure when accessing types from dependencies compiled with newer JDK versions that use APIs unavailable in the current runtime.

For example, when a library compiled with JDK 21 (using
`java.util.concurrent.StructuredTaskScope`) is used in a project running on JDK 17, the compiler currently crashes with an unhandled `MissingType` exception during the Erasure phase instead of reporting an error.

`MissingType` is thrown from `TypeErasure.sigName` or `checkedSuperType`, which are called from `TypeErasure.eraseXXX`.
These errors are usually caught by `Typer.handleTypeError` or `TreePickler.pickleType`, but those handlers do not cover exceptions thrown during the Erasure. (It seems that during typer, the compiler trusts the information from the classfile or tasty without verifying that all referenced types actually exist on the classpath).

This commit encloses `Erasure.transform` with a try-catch to intercept the `TypeError` and call `report.error`.