Also substitute case bindings in symbols
The fix for #24813 is the also updating the symbol in the line
```
to.foreach(sym => sym.info = sym.info.substSym(from, to))
```

the other changes are more clean up so we only have to do that
substitution once. Before https://github.com/scala/scala3/pull/7902
the substitution are done both inside substBindings and afterwards. But
if done after it the partial substitutions inside are redundant. So this
changes it so we only call `subst` one place.

It unclear if we expect `subst` to already do this replace? If we do
then we should probably fix this in that implementation instead.