Warn if reach capability in result will cause leakage
An unboxed reach in a result type of a method will in almost all cases lead
to use leakage error in the body of the method (except if the body is pure, and
we widen to the reach without cause). Warn at this point already so that the
leakage error that follows can be better understood.